+++ pkarr/src/client/mod.rs
@@ -5,7 +5,7 @@ pub mod cache;
 #[cfg(all(not(target_arch = "wasm32"), feature = "dht"))]
 pub(crate) mod dht;
 #[cfg(all(not(target_arch = "wasm32"), feature = "dht"))]
-pub use dht::{Client, Config};
+pub use dht::{resolvres_to_socket_addrs, Client, Config};
 
 #[cfg(target_arch = "wasm32")]
 pub(crate) mod relay;