+++ pkarr/CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to pkarr will be documented in this file.
 - Add feature `reqwest-builder` to create a `reqwest::ClientBuilder` from `Client` and `relay::client::Client` using custom dns resolver and preconfigured rustls client config.
 - Implement `FromStr` for `PublicKey`
 - Implement `TryFrom<MutableItem>` for `SignedPacket`
+- Add `resolvres_to_socket_addrs()` function.
 
 ### Changed
 
@@ -35,6 +36,8 @@ All notable changes to pkarr will be documented in this file.
 - `Client::resolve`, `Client::resolve_sync` and `relay::Client::resolve` return expired cached `SignedPacket` _before_ making query to the network (Relays/Resolvers/Dht).
 - Update `simple-dns` so you can't use `Name::new("@")`, instead you should use `Name::new(".")`, `SignedPacket::resource_records("@")` still works.
 - Replace `ClientBuilder::testnet()` with `ClientBuilder::bootstrap()`.
+- Change `SignedPacket::to_relay_payload()` to `SignedPacket::as_relay_payload()`.
+- Replace `bytes::Bytes` return types with `&[u8]`.
 
 ### Removed
 
@@ -42,3 +45,4 @@ All notable changes to pkarr will be documented in this file.
 - Removed `SignedPacket::from_packet`.
 - Removed `SignedPacket::packet` getter.
 - Removed rexported `mainline`
+- Remove `bytes` dependency.