+++ relay/CHANGELOG.md
@@ -6,4 +6,5 @@ All notable changes to pkarr relay will be documented in this file.
 
 ### Added
 
-- Add `Relay` and `Config` to be able run an in-process relay with full control over configurations
+- Add `Relay` and `Config` to be able run an in-process relay with full control over configurations.
+- Use `LmdbCache` from `pkarr`.